backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.3.0
pyjwt==2.8.0
cryptography==41.0.7
psycopg2-binary==2.9.9

